Added boilerplate and lib dir
lib dir is used to store all common libraries that would require
agnostism e.g. platform
lib dir is also used to store libraries that are commonly used within
the source code
The purpose of this lib dir is to ensure that the Lingua code only
contains code that is focusing on the controls of the language and
detaching from the controls of other systems

Third-party system controls are stored in lib dir to ensure that Lingua
works on any platform

Task dir is used to link between the REST framework to the 3rd party
system.

Signed-off-by: Stanley Phoong <stanley.phoong@gmail.com>